Fix nativesubtitles for iOS 18
iOS 18 seems to fail to play a video if it has crossorigin="anonymous".
At the same time it does not have a problem to load VTT subtitles from other domains, even without CORS headers.
Same applies to iOS 17 - except it does not fail to play a video if it has crossorigin="anonymous".